refactor: extract sessions into sessions-rds.ts
